+++ lib/pghero/database.rb
@@ -118,12 +118,6 @@ module PgHero
       @filter_data
     end
 
-    # TODO remove in next major version
-    alias_method :access_key_id, :aws_access_key_id
-    alias_method :secret_access_key, :aws_secret_access_key
-    alias_method :region, :aws_region
-    alias_method :db_instance_identifier, :aws_db_instance_identifier
-
     private
 
     # check adapter lazily